+++ examples/with-mobx-state-tree/pages/_app.js
@@ -1,43 +1,12 @@
-import React from 'react'
 import { Provider } from 'mobx-react'
-import { getSnapshot } from 'mobx-state-tree'
-import App from 'next/app'
-import { initializeStore } from '../stores/store'
+import { useStore } from '../store'
 
-export default class MyApp extends App {
-  static async getInitialProps({ Component, ctx }) {
-    //
-    // Use getInitialProps as a step in the lifecycle when
-    // we can initialize our store
-    //
-    const isServer = typeof window === 'undefined'
-    const store = initializeStore(isServer)
-    //
-    // Check whether the page being rendered by the App has a
-    // static getInitialProps method and if so call it
-    //
-    let pageProps = {}
-    if (Component.getInitialProps) {
-      pageProps = await Component.getInitialProps(ctx)
-    }
-    return {
-      initialState: getSnapshot(store),
-      isServer,
-      pageProps,
-    }
-  }
+export default function App({ Component, pageProps }) {
+  const store = useStore(pageProps.initialState)
 
-  constructor(props) {
-    super(props)
-    this.store = initializeStore(props.isServer, props.initialState)
-  }
-
-  render() {
-    const { Component, pageProps } = this.props
-    return (
-      <Provider store={this.store}>
-        <Component {...pageProps} />
-      </Provider>
-    )
-  }
+  return (
+    <Provider store={store}>
+      <Component {...pageProps} />
+    </Provider>
+  )
 }
