+++ examples/with-mobx-state-tree/pages/index.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import SampleComponent from '../components/SampleComponent'
 
 export default () => {
