+++ examples/with-mobx-state-tree/pages/other.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import SampleComponent from '../components/SampleComponent'
 
 export default () => {
