+++ examples/with-mobx-state-tree/pages/ssg.js
@@ -1,0 +1,17 @@
+import { getSnapshot } from 'mobx-state-tree'
+import SampleComponent from '../components/SampleComponent'
+import { initializeStore } from '../store'
+
+export default () => {
+  return <SampleComponent title={'SSG Page'} linkTo="/" />
+}
+
+// If you build and start the app, the date returned here will have the same
+// value for all requests, as this method gets executed at build time.
+export function getStaticProps() {
+  const store = initializeStore()
+
+  store.update()
+
+  return { props: { initialState: getSnapshot(store) } }
+}
